Tracking gov websites creation
First commit to this project. This script pulls the domains from CISAGov that are registered to the US Government and runs them all through WHOIS to augment the data with registration information. The data allows us to track the creation of websites over time. The code than generates the GIF tracking the counts by administration over time.